+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@athenna/cli",
-  "version": "4.0.0",
+  "version": "4.1.0",
   "description": "Athenna CLI to create new Athenna projects.",
   "license": "MIT",
   "author": "João Lenon <lenon@athenna.io>",
@@ -48,18 +48,17 @@
     "#tests": "./tests/index.js"
   },
   "dependencies": {
-    "@athenna/artisan": "^4.5.0",
-    "@athenna/common": "^4.9.1",
-    "@athenna/config": "^4.3.0",
-    "@athenna/core": "^4.4.0",
-    "@athenna/ioc": "^4.1.0",
-    "@athenna/logger": "^4.2.0",
-    "@athenna/view": "^4.1.0",
-    "cross-env": "^7.0.3",
+    "@athenna/artisan": "^4.10.0",
+    "@athenna/common": "^4.13.2",
+    "@athenna/config": "^4.4.0",
+    "@athenna/core": "^4.9.1",
+    "@athenna/ioc": "^4.4.1",
+    "@athenna/logger": "^4.5.0",
+    "@athenna/view": "^4.3.0",
     "reflect-metadata": "^0.1.13"
   },
   "devDependencies": {
-    "@athenna/test": "^4.3.0",
+    "@athenna/test": "^4.7.0",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
     "c8": "^7.12.0",
